Add job for releasing macOS universal executable
The workflow is based swift-workflow-test. Because this is a proof of
concept, I removed the jobs for Linux and Windows.

The work dispatch input was not used in the workflow. So, I removed it.

The conditional on release step was removed because it is already
covered by `on push`.

The build step was adjusted since the release worklow is intended for
usage in swift-markdown-babel.

See also: https://github.com/Optiroc/swift-workflow-test/blob/main/.github/workflows/release.yml
See also: https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#onworkflow_dispatch